feature: reorganize handlers, fix telegram linking docs, remove redundant env copy in Dockerfile, write air config for hot reload bot development,  set long polling for testing, cahnge go version to 1.24.4
