--- backend/.env.example
@@ -1,50 +1,0 @@
-# App Configuration
-APP_NAME=FluentlyAPI
-APP_HOST=0.0.0.0
-APP_PORT=8070
-
-# Database Configuration
-DB_HOST=localhost
-DB_PORT=5432
-DB_USER=postgres
-DB_PASSWORD=your_password
-DB_NAME=fluently
-
-# JWT Configuration
-JWT_SECRET=your_super_secret_jwt_key_here_make_it_long_and_secure
-JWT_EXPIRATION=24h
-REFRESH_EXPIRATION=720h
-
-# Google OAuth Configuration
-IOS_GOOGLE_CLIENT_ID=your_ios_client_id
-ANDROID_GOOGLE_CLIENT_ID=your_android_client_id
-WEB_GOOGLE_CLIENT_ID=your_web_client_id
-WEB_GOOGLE_CLIENT_SECRET=your_web_client_secret
-
-# Logging Configuration
-LOG_LEVEL=info
-LOG_PATH=./logs/app.log
-
-# Security Configuration
-PASSWORD_MIN_LENGTH=8
-RATE_LIMIT_REQUESTS=100
-RATE_LIMIT_DURATION=1h
-
-# Swagger Configuration (comma-separated emails)
-SWAGGER_ALLOWED_EMAILS=admin@example.com,developer@example.com
-SWAGGER_HOST=10.243.92.227:8070 
-
-# Directus Configuration
-DIRECTUS_SECRET_KEY=your_directus_secret_key_here_make_it_long_and_secure
-DIRECTUS_ADMIN_EMAIL=admin@fluently-app.ru
-DIRECTUS_ADMIN_PASSWORD=your_directus_admin_password_here
-DIRECTUS_PORT=8055
-
-# Grafana Configuration
-GRAFANA_ADMIN_PASSWORD=your_grafana_admin_password_here
-
-# ZeroTier Configuration
-ZEROTIER_IP=10.243.92.227
-
-# Public URLs
-PUBLIC_URL=https://fluently-app.ru
