--- telegram-bot/.env.example
@@ -1,38 +1,0 @@
-# Telegram Bot Configuration
-BOT_TOKEN=
-WEBHOOK_URL=
-WEBHOOK_PORT=8080
-
-# Webhook Configuration  
-WEBHOOK_HOST=0.0.0.0
-WEBHOOK_PATH=/webhook
-WEBHOOK_SECRET=
-WEBHOOK_CERT_FILE=
-WEBHOOK_KEY_FILE=
-WEBHOOK_MAX_BODY_SIZE=1048576
-
-# Domain Configuration (for NGINX template)
-DOMAIN=
-CERT_NAME=
-
-# Redis Configuration
-REDIS_ADDR=localhost:6379
-REDIS_PASSWORD=
-REDIS_DB=0
-
-# Backend API Configuration
-API_BASE_URL=
-API_TIMEOUT=30
-
-# Asynq Configuration (for background tasks)
-ASYNQ_REDIS_ADDR=localhost:6379
-ASYNQ_REDIS_PASSWORD=
-ASYNQ_REDIS_DB=1
-ASYNQ_CONCURRENCY=10
-
-# Logging Configuration
-LOG_LEVEL=info
-LOG_PATH=logs/bot.log
-
-# Environment
-ENVIRONMENT=development
