+++ telegram-bot/go.mod
@@ -1,20 +1,15 @@
 // go.mod
 module telegram-bot
 
-go 1.22
-
-toolchain go1.24.4
+go 1.24.4
 
 require (
-	github.com/go-chi/chi/v5 v5.2.2
-	github.com/go-chi/cors v1.2.2
 	github.com/google/uuid v1.6.0
 	github.com/hibiken/asynq v0.25.1
 	github.com/joho/godotenv v1.5.1
 	github.com/redis/go-redis/v9 v9.7.0
 	github.com/spf13/viper v1.16.0
 	go.uber.org/zap v1.21.0
-	golang.org/x/time v0.8.0
 	gopkg.in/telebot.v3 v3.3.8
 )
 
@@ -36,6 +31,7 @@ require (
 	go.uber.org/multierr v1.8.0 // indirect
 	golang.org/x/sys v0.27.0 // indirect
 	golang.org/x/text v0.9.0 // indirect
+	golang.org/x/time v0.8.0 // indirect
 	google.golang.org/protobuf v1.35.2 // indirect
 	gopkg.in/ini.v1 v1.67.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
