+++ telegram-bot/internal/bot/handlers/help.go
@@ -9,28 +9,28 @@ import (
 func HelpHandler(c tele.Context) error {
 	text := `*Справка по командам бота*
 
-/start - Начало работы с ботом
-/learn - Начать изучение новых слов
-/settings - Настройки обучения
-/help - Показать эту справку
+		/start - Начало работы с ботом
+		/learn - Начать изучение новых слов
+		/settings - Настройки обучения
+		/help - Показать эту справку
 
-*О боте*
-Fluently - приложение для изучения английского языка на основе научно обоснованной методики.
+		*О боте*
+		Fluently - приложение для изучения английского языка на основе научно обоснованной методики.
 
-Основные особенности:
-• Фокус на самых важных словах (80-90% речи)
-• Особая техника запоминания
-• Всего 10 слов в день
-• Оптимальная интенсивность обучения
+		Основные особенности:
+		• Фокус на самых важных словах (80-90% речи)
+		• Особая техника запоминания
+		• Всего 10 слов в день
+		• Оптимальная интенсивность обучения
 
-*Как пользоваться*
-1. Настройте количество слов в день в разделе Настройки
-2. Включите уведомления, чтобы не пропустить ежедневные занятия
-3. Пройдите тест для определения вашего уровня
-4. Занимайтесь по 10-15 минут каждый день
-5. Отслеживайте свой прогресс
+		*Как пользоваться*
+		1. Настройте количество слов в день в разделе Настройки
+		2. Включите уведомления, чтобы не пропустить ежедневные занятия
+		3. Пройдите тест для определения вашего уровня
+		4. Занимайтесь по 10-15 минут каждый день
+		5. Отслеживайте свой прогресс
 
-Если у вас возникли вопросы или проблемы, напишите в поддержку: support@fluently.app`
+		Если у вас возникли вопросы или проблемы, напишите в поддержку: support@fluently.app`
 
 	return c.Send(text, &tele.SendOptions{ParseMode: tele.ModeMarkdown})
 }
