+++ telegram-bot/internal/bot/handlers/learning.go
@@ -20,7 +20,6 @@ func (s *HandlerService) HandleLearningStart(ctx context.Context, c tele.Context
 			"id":          int64(1),
 			"word":        "hello",
 			"translation": "привет",
-			"definition":  "используется как приветствие",
 			"examples":    []string{"Hello, how are you? — Привет, как дела?", "She said hello to everyone. — Она поприветствовала всех."},
 			"cefr_level":  "A1",
 		},
@@ -28,7 +27,6 @@ func (s *HandlerService) HandleLearningStart(ctx context.Context, c tele.Context
 			"id":          int64(2),
 			"word":        "world",
 			"translation": "мир",
-			"definition":  "земля и все люди, места и вещи на ней",
 			"examples":    []string{"The world is beautiful. — Мир прекрасен.", "People from around the world came to visit. — Люди со всего мира приехали в гости."},
 			"cefr_level":  "A1",
 		},
@@ -36,7 +34,6 @@ func (s *HandlerService) HandleLearningStart(ctx context.Context, c tele.Context
 			"id":          int64(3),
 			"word":        "beautiful",
 			"translation": "красивый",
-			"definition":  "обладающий красотой; приятный для чувств",
 			"examples":    []string{"She is beautiful. — Она красивая.", "What a beautiful day! — Какой прекрасный день!"},
 			"cefr_level":  "A2",
 		},
